+++ game_patch/main.cpp
@@ -206,6 +206,7 @@ FunHook<void(bool)> LevelInitPost_hook{
     0x00435DF0,
     [](bool is_auto_level_load) {
         LevelInitPost_hook.CallTarget(is_auto_level_load);
+        GraphicsLevelInitPost();
         xlog::info("Level loaded: %s%s", rf::level_filename.CStr(), is_auto_level_load ? " (caused by event)" : "");
     },
 };